Merge geographic codes

Convert geographic codes while preserving the number of rows in a data frame or the dimension of a vector. Merging is handled, but not splitting.

Usage

codes_to_one(
  data,
  codes_ini,
  codes_fin,
  extra = NULL,
  from = NULL,
  to = NULL,
  ...
)

Arguments

data

Data frame or vector

codes_ini

Initial geographic codes

codes_fin

Final geographic codes

extra

Other geographic codes

from

Initial column (in a data frame)

to

Final column (in a data frame)

...

Arguments passed to or from other methods.

Value

a data frame with the same number of rows or a vector of equal dimension

Examples

x <- c("B","C1","C2","A","Z")
codes_1 <- c("A","A","B","C1","C2")
codes_2 <-  c("a1","a2","b","c","c")

# Data frame
dt <- data.frame(X = x)
codes_to_one(dt, from = "X", to = "Y", codes_ini = codes_1, codes_fin = codes_2)
#>    X    Y
#> 1  B    b
#> 2 C1    c
#> 3 C2    c
#> 4  A   a1
#> 5  Z <NA>

# Vector
codes_to_one(x, codes_ini = codes_1, codes_fin = codes_2)
#> [1] "b"  "c"  "c"  "a1" NA